requirements.txt
EbookLib~=0.17.1
PyYAML~=5.4.1
nltk~=3.6.2
tika~=1.24
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests~=2.26.0